requirements.txt
tensorflow-cpu
numpy
matplotlib
scikit-learn
seaborn